requirements-dev.txt
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
d==0.2.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.1
django-cors==0.1
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
heroku==0.1.4
idna==2.8
lxml==4.3.3
Markdown==3.1
Pillow==6.0.0
psycopg2==2.7.7
Pygments==2.4.0
pyquery==1.4.0
python-dateutil==1.5
python-decouple==3.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
